# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UrjaKavach API",
    "description": "Optional demo API. Simulated artifacts, assumption-driven estimates and local demo roles. No production authentication.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/assets/{asset_id}": {
      "get": {
        "summary": "Asset",
        "operationId": "asset_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/scenarios/{scenario_id}/telemetry": {
      "get": {
        "summary": "Telemetry",
        "operationId": "telemetry_scenarios__scenario_id__telemetry_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "signals",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "gearbox_temperature,expected_temperature",
              "title": "Signals"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 10000,
              "title": "To"
            }
          },
          {
            "name": "max_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 3,
              "default": 200,
              "title": "Max Points"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenario_id}/scores": {
      "get": {
        "summary": "Scores",
        "operationId": "scores_scenarios__scenario_id__scores_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "B0",
                "M1",
                "M2",
                "M3"
              ],
              "type": "string",
              "default": "M2",
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "Alerts",
        "operationId": "alerts_alerts_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts/{alert_id}/ack": {
      "post": {
        "summary": "Acknowledge",
        "operationId": "acknowledge_alerts__alert_id__ack_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts/{alert_id}/evidence": {
      "get": {
        "summary": "Evidence",
        "operationId": "evidence_alerts__alert_id__evidence_get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "get": {
        "summary": "Compare",
        "operationId": "compare_compare_get",
        "parameters": [
          {
            "name": "scenario_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario A"
            }
          },
          {
            "name": "scenario_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario B"
            }
          },
          {
            "name": "signal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "gearbox_temperature",
              "title": "Signal"
            }
          },
          {
            "name": "window_a",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0,287",
              "title": "Window A"
            }
          },
          {
            "name": "window_b",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0,287",
              "title": "Window B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/assumptions": {
      "get": {
        "summary": "Assumptions",
        "operationId": "assumptions_assumptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Save Assumptions",
        "operationId": "save_assumptions_assumptions_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssumptionValues"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/loss/estimate": {
      "post": {
        "summary": "Loss",
        "operationId": "loss_loss_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LossRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/optimize": {
      "post": {
        "summary": "Schedule",
        "operationId": "schedule_schedule_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedules/{schedule_id}": {
      "get": {
        "summary": "Get Schedule",
        "operationId": "get_schedule_schedules__schedule_id__get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LossRequest": {
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/AssumptionValues"
          },
          "delay_days": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 0.0,
            "title": "Delay Days",
            "default": 3
          },
          "horizon_days": {
            "type": "integer",
            "maximum": 90.0,
            "minimum": 1.0,
            "title": "Horizon Days",
            "default": 14
          },
          "draws": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 100.0,
            "title": "Draws",
            "default": 1000
          },
          "seed": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 2026
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LossRequest"
      },
...
    }
  }
}